starting to fit multiple sightlines
